Types conversion handling cleanup
This PR contains chore changes with some straightening aroung handling
of the conversion between MySQL and DuckDB types.

Testing: no functional changes, no new tests.
