+++ src/include/mysql_utils.hpp
@@ -15,21 +15,6 @@ namespace duckdb {
 class MySQLSchemaEntry;
 class MySQLTransaction;
 
-struct MySQLTypeData {
-	string type_name;
-	string column_type;
-	int64_t precision;
-	int64_t scale;
-};
-
-enum class MySQLTypeAnnotation { STANDARD, CAST_TO_VARCHAR, NUMERIC_AS_DOUBLE, CTID, JSONB, FIXED_LENGTH_CHAR };
-
-struct MySQLType {
-	idx_t oid = 0;
-	MySQLTypeAnnotation info = MySQLTypeAnnotation::STANDARD;
-	vector<MySQLType> children;
-};
-
 struct MySQLConnectionParameters {
 	string host;
 	string user;
@@ -54,11 +39,6 @@ class MySQLUtils {
 public:
 	static std::tuple<MySQLConnectionParameters, unordered_set<string>> ParseConnectionParameters(const string &dsn);
 	static MYSQL *Connect(const string &dsn);
-
-	static LogicalType ToMySQLType(ClientContext &context, const LogicalType &input);
-	static LogicalType TypeToLogicalType(ClientContext &context, const MySQLTypeData &input);
-	static LogicalType FieldToLogicalType(ClientContext &context, MYSQL_FIELD *field);
-	static string TypeToString(const LogicalType &input);
 
 	static string WriteIdentifier(const string &identifier);
 	static string WriteLiteral(const string &identifier);
