+++ src/storage/CMakeLists.txt
@@ -9,7 +9,6 @@ add_library(
   mysql_index_set.cpp
   mysql_insert.cpp
   mysql_optimizer.cpp
-  mysql_result.cpp
   mysql_schema_entry.cpp
   mysql_schema_set.cpp
   mysql_table_entry.cpp
